Some folders weren't uploaded on the first go.
This commit adds the "build" folder, but apparently the "bin" folder
isn't capable of being added, so I've added a README.txt to describe how
to create it. It's possible to build the project from the contents of
the working folder without a functioning copy of lpp.exe, but it's
better to provide both avenues.